+++ cs/example/Samples/Custom.cs
@@ -4,7 +4,7 @@
  * Created Date: 14/10/2021
  * Author: Shun Suzuki
  * -----
- * Last Modified: 21/10/2022
+ * Last Modified: 07/11/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -32,8 +32,8 @@ internal static class CustomTest
             {
                 var tp = tr.Position;
                 var dist = (tp - point).L2Norm;
-                var wavelength = tr.Wavelength;
-                var phase = dist / wavelength;
+                var wavenum = 2.0 * AUTD3.Pi / tr.Wavelength;
+                var phase = dist * wavenum;
                 amps[tr.Id] = 1.0;
                 phases[tr.Id] = phase;
             }