+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.5.2
 
 - Fix [#37](https://github.com/shinolab/autd3/issues/37)
+- Change phase unit to radian
 
 # 2.5.1